pyproject.toml
[tool.poetry]
name = "dumb_smart_home"
version = "0.1.0"
description = ""
authors = ["Jon Gaul <jgaul3@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.1"
uvicorn = "^0.22.0"
jinja2 = "^3.1.2"
python-multipart = "^0.0.6"
pyjwt = "^2.6.0"
requests = "^2.30.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
